+++ packages/block-library/src/cover/overlay-color-settings.native.js
@@ -12,9 +12,9 @@ import {
 	getGradientValueBySlug,
 	getGradientSlugByValue,
 	__experimentalPanelColorGradientSettings as PanelColorGradientSettings,
-	useSetting,
 } from '@wordpress/block-editor';
 import { useMemo } from '@wordpress/element';
+import { useMobileGlobalStylesColors } from '@wordpress/components';
 
 function OverlayColorSettings( {
 	overlayColor,
@@ -23,9 +23,8 @@ function OverlayColorSettings( {
 	customGradient,
 	setAttributes,
 } ) {
-	const EMPTY_ARRAY = [];
-	const colors = useSetting( 'color.palette' ) || EMPTY_ARRAY;
-	const gradients = useSetting( 'color.gradients' ) || EMPTY_ARRAY;
+	const colors = useMobileGlobalStylesColors();
+	const gradients = useMobileGlobalStylesColors( 'gradients' );
 
 	const gradientValue =
 		customGradient || getGradientValueBySlug( gradients, gradient );